Fixed bug with moving into tokens as tokens generated on gameplay level. Fixed logic with checking movements before combat
