+++ GamePlay.tscn
@@ -1058,12 +1058,9 @@ pattern_24 = SubResource("TileMapPattern_3knyh")
 pattern_25 = SubResource("TileMapPattern_hai2l")
 pattern_26 = SubResource("TileMapPattern_nxmvw")
 
-[node name="GamePlay" type="Node2D" node_paths=PackedStringArray("mapGen", "deck", "player")]
+[node name="GamePlay" type="Node2D"]
 z_index = -1
 script = ExtResource("1_m0gf8")
-mapGen = NodePath("GameplayControl/MapGen")
-deck = NodePath("GameplayControl/PlayerUI/PlayerArea/Deck")
-player = NodePath("GameplayControl/Player")
 
 [node name="GameplayControl" type="Control" parent="." node_paths=PackedStringArray("mapGen", "player", "gamePlay")]
 layout_mode = 3
@@ -1074,6 +1071,7 @@ offset_right = 1920.0
 offset_bottom = 1080.0
 grow_horizontal = 2
 grow_vertical = 2
+mouse_filter = 1
 script = ExtResource("2_7jlcl")
 mapGen = NodePath("MapGen")
 player = NodePath("Player")
@@ -1298,13 +1296,13 @@ offset_bottom = 147.0
 text = "Undo"
 metadata/_edit_use_anchors_ = true
 
-[node name="MapGen" type="TileMap" parent="GameplayControl" node_paths=PackedStringArray("gamePlay")]
+[node name="MapGen" type="TileMap" parent="GameplayControl" node_paths=PackedStringArray("gameplayControl")]
 position = Vector2(482, 912)
 tile_set = SubResource("TileSet_cu01s")
 format = 2
 layer_0/tile_data = PackedInt32Array(65535, 131072, 0, 131071, 327680, 0, 65536, 393216, 0, -65535, 65536, 0, 1, 262144, 0, -65536, 0, 0, 0, 196608, 0)
 script = ExtResource("3_hiwow")
-gamePlay = NodePath("../..")
+gameplayControl = NodePath("..")
 
 [node name="Camera2D" type="Camera2D" parent="GameplayControl"]
 position = Vector2(969, 543)
