+++ MapGen/MapGen.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 public partial class MapGen : TileMap
 {
 	[Export]
-	private GamePlay gamePlay;
+	private GameplayControl gameplayControl;
 	const int MainLayer = 0;
 	const int MainAtlasID = 0;
 	const int MainTerrainSet = 0;
@@ -95,7 +95,7 @@ public partial class MapGen : TileMap
 					{
 						// May need to add in switch statement for whether token is flipped
 						// Generate monster from color stack, site fortifications from what site it's on, on what tile
-						gamePlay.MonsterGen(tokenData, (patternTileData.GetCustomData("Event").ToString() == "") ? 0 : 1, patternTile);
+						gameplayControl.MonsterGen(tokenData, (patternTileData.GetCustomData("Event").ToString() == "") ? 0 : 1, patternTile);
 					}
 				}
